Extract DefaultToolFiltersStateProducer from ToolsPresenter
