+++ app/src/main/kotlin/org/cru/godtools/ui/dashboard/tools/ToolsModule.kt
@@ -1,0 +1,13 @@
+package org.cru.godtools.ui.dashboard.tools
+
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal abstract class ToolsModule {
+    @Binds
+    abstract fun toolFiltersStateProducer(impl: DefaultToolFiltersStateProducer): ToolFiltersStateProducer
+}
