+++ app/src/testDebug/kotlin/org/cru/godtools/ui/dashboard/tools/ToolsLayoutPaparazziTest.kt
@@ -29,6 +29,7 @@ import org.cru.godtools.model.Language
 import org.cru.godtools.model.Tool
 import org.cru.godtools.model.randomTool
 import org.cru.godtools.ui.dashboard.filters.FilterMenu
+import org.cru.godtools.ui.dashboard.tools.ToolFiltersStateProducer.Filters
 import org.cru.godtools.ui.dashboard.tools.ToolsPresenter.UiState
 import org.cru.godtools.ui.tools.ToolCardStateTestData
 import org.junit.Assume.assumeTrue
@@ -104,7 +105,7 @@ class ToolsLayoutPaparazziTest(
     @Test
     fun `ToolsLayout() - Filters Selected`() = snapshotToolsLayout(
         state.copy(
-            filters = UiState.Filters(
+            filters = Filters(
                 categoryFilter = FilterMenu.UiState(selectedItem = Tool.CATEGORY_GOSPEL),
                 languageFilter = FilterMenu.UiState(
                     selectedItem = Language(Locale.ENGLISH),
@@ -118,7 +119,7 @@ class ToolsLayoutPaparazziTest(
     @Ignore("LayoutLib does not correctly support Popups/Windows currently")
     fun `ToolsLayout() - Language Filter Expanded`() = snapshotToolsLayout(
         state.copy(
-            filters = UiState.Filters(
+            filters = Filters(
                 languageFilter = FilterMenu.UiState(
                     selectedItem = Language(Locale.ENGLISH),
                     menuExpanded = mutableStateOf(true),
