+++ app/src/testDebug/kotlin/org/cru/godtools/ui/dashboard/tools/ToolsPresenterTest.kt
@@ -1,39 +1,30 @@
 package org.cru.godtools.ui.dashboard.tools
 
 import android.app.Application
-import androidx.test.core.app.ApplicationProvider
+import androidx.compose.runtime.mutableStateOf
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.jeppeman.mockposable.mockk.everyComposable
 import com.slack.circuit.test.FakeNavigator
 import com.slack.circuit.test.test
-import io.mockk.coEvery
 import io.mockk.every
-import io.mockk.excludeRecords
 import io.mockk.mockk
-import io.mockk.verify
-import io.mockk.verifyAll
 import java.util.Locale
+import kotlin.random.Random
 import kotlin.test.AfterTest
-import kotlin.test.BeforeTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
+import kotlin.uuid.ExperimentalUuidApi
+import kotlin.uuid.Uuid
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.test.TestScope
-import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
 import org.ccci.gto.android.common.androidx.compose.ui.platform.AndroidUiDispatcherUtil
 import org.cru.godtools.base.Settings
-import org.cru.godtools.db.repository.LanguagesRepository
 import org.cru.godtools.db.repository.ToolsRepository
-import org.cru.godtools.db.repository.TranslationsRepository
 import org.cru.godtools.model.Language
 import org.cru.godtools.model.Tool
-import org.cru.godtools.model.Translation
 import org.cru.godtools.model.randomTool
-import org.cru.godtools.model.randomTranslation
 import org.cru.godtools.ui.banner.BannerType
 import org.cru.godtools.ui.dashboard.filters.FilterMenu
 import org.cru.godtools.ui.tools.ToolCard
@@ -43,76 +34,45 @@ import org.robolectric.annotation.Config
 
 @RunWith(AndroidJUnit4::class)
 @Config(application = Application::class)
-@Suppress("UnusedFlow")
 @OptIn(ExperimentalCoroutinesApi::class)
 class ToolsPresenterTest {
-    private val appLanguage = MutableStateFlow(Locale.ENGLISH)
     private val isFavoritesFeatureDiscovered = MutableStateFlow(true)
     private val toolsFlow = MutableStateFlow(emptyList<Tool>())
     private val filteredToolsFlow = MutableStateFlow(emptyList<Tool>())
-    private val languagesFlow = MutableStateFlow(emptyList<Language>())
-    private val gospelLanguagesFlow = MutableStateFlow(emptyList<Language>())
-    private val selectedCategory = MutableStateFlow<String?>(null)
-    private val selectedLocale = MutableStateFlow<Locale?>(null)
-
-    private val testScope = TestScope()
 
     private val filteredToolsFlowProducer: FilteredToolsFlowProducer = mockk {
         every { getFlow(any(), any()) } returns filteredToolsFlow
     }
-    private val languagesRepository: LanguagesRepository = mockk {
-        every { findLanguageFlow(any()) } returns flowOf(null)
-        every { getLanguagesFlow() } returns languagesFlow
-        every { getLanguagesFlowForToolCategory(Tool.CATEGORY_GOSPEL) } returns gospelLanguagesFlow
-
-        excludeRecords { this@mockk.equals(any()) }
-    }
     private val navigator = FakeNavigator(ToolsScreen)
     private val settings: Settings = mockk {
-        every { appLanguage } returns this@ToolsPresenterTest.appLanguage.value
-        every { appLanguageFlow } returns this@ToolsPresenterTest.appLanguage
         every { isFeatureDiscoveredFlow(Settings.FEATURE_TOOL_FAVORITE) } returns isFavoritesFeatureDiscovered
-        every { getDashboardFilterCategoryFlow() } returns selectedCategory
-        every { getDashboardFilterLocaleFlow() } returns selectedLocale
-        coEvery { updateDashboardFilterCategory(any()) } answers { selectedCategory.value = firstArg() }
-        coEvery { updateDashboardFilterLocale(any()) } answers { selectedLocale.value = firstArg() }
     }
     private val toolsRepository: ToolsRepository = mockk {
         every { getNormalToolsFlow() } returns toolsFlow
     }
-    private val translationsRepository: TranslationsRepository = mockk {
-        every { getTranslationsFlowForTools(any()) } returns flowOf(emptyList())
-    }
+    private val toolFiltersStateProducer = FakeToolFiltersStateProducer()
 
     private val toolCardPresenter: ToolCardPresenter = mockk {
         everyComposable { present(tool = any(), secondLanguage = any(), eventSink = any()) }
             .answers { ToolCard.State(tool = firstArg()) }
     }
 
-    private lateinit var presenter: ToolsPresenter
-
-    @BeforeTest
-    fun setup() {
-        presenter = ToolsPresenter(
-            context = ApplicationProvider.getApplicationContext(),
-            eventBus = mockk(),
-            settings = settings,
-            toolCardPresenter = toolCardPresenter,
-            languagesRepository = languagesRepository,
-            toolsRepository = toolsRepository,
-            translationsRepository = translationsRepository,
-            filteredToolsFlowProducer = filteredToolsFlowProducer,
-            ioDispatcher = UnconfinedTestDispatcher(testScope.testScheduler),
-            navigator = navigator,
-        )
-    }
+    private val presenter = ToolsPresenter(
+        eventBus = mockk(),
+        settings = settings,
+        toolCardPresenter = toolCardPresenter,
+        toolsRepository = toolsRepository,
+        filteredToolsFlowProducer = filteredToolsFlowProducer,
+        toolFiltersStateProducer = toolFiltersStateProducer,
+        navigator = navigator,
+    )
 
     @AfterTest
     fun cleanup() = AndroidUiDispatcherUtil.runScheduledDispatches()
 
     // region State.banner
     @Test
-    fun `State - banner - none`() = testScope.runTest {
+    fun `State - banner - none`() = runTest {
         presenter.test {
             isFavoritesFeatureDiscovered.value = true
             assertNull(expectMostRecentItem().banner)
@@ -120,7 +80,7 @@ class ToolsPresenterTest {
     }
 
     @Test
-    fun `State - banner - favorites`() = testScope.runTest {
+    fun `State - banner - favorites`() = runTest {
         presenter.test {
             isFavoritesFeatureDiscovered.value = false
             assertEquals(BannerType.TOOL_LIST_FAVORITES, expectMostRecentItem().banner)
@@ -130,7 +90,7 @@ class ToolsPresenterTest {
 
     // region State.spotlightTools
     @Test
-    fun `Property spotlightTools`() = testScope.runTest {
+    fun `Property spotlightTools`() = runTest {
         val normalTool = randomTool("normal", isHidden = false, isSpotlight = false)
         val spotlightTool = randomTool("spotlight", isHidden = false, isSpotlight = true)
 
@@ -141,7 +101,7 @@ class ToolsPresenterTest {
     }
 
     @Test
-    fun `Property spotlightTools - Don't show hidden tools`() = testScope.runTest {
+    fun `Property spotlightTools - Don't show hidden tools`() = runTest {
         val hiddenTool = randomTool("normal", isHidden = true, isSpotlight = true)
         val spotlightTool = randomTool("spotlight", isHidden = false, isSpotlight = true)
 
@@ -152,7 +112,7 @@ class ToolsPresenterTest {
     }
 
     @Test
-    fun `Property spotlightTools - Sorted by default order`() = testScope.runTest {
+    fun `Property spotlightTools - Sorted by default order`() = runTest {
         val tools = List(10) {
             randomTool("tool$it", Tool.Type.TRACT, defaultOrder = it, isHidden = false, isSpotlight = true)
         }
@@ -164,192 +124,39 @@ class ToolsPresenterTest {
     }
     // endregion State.spotlightTools
 
-    // region State.filters.categoryFilter.items
+    // region State.filters
     @Test
-    fun `State - filters - categoryFilter - items - distinct & ordered by tool order`() = testScope.runTest {
-        filteredToolsFlow.value = listOf(
-            randomTool(category = Tool.CATEGORY_ARTICLES),
-            randomTool(category = Tool.CATEGORY_GOSPEL),
-            randomTool(category = Tool.CATEGORY_ARTICLES),
+    @OptIn(ExperimentalUuidApi::class)
+    fun `State - filters`() = runTest {
+        val filters = ToolFiltersStateProducer.Filters(
+            categoryFilter = FilterMenu.UiState(
+                menuExpanded = mutableStateOf(Random.nextBoolean()),
+                items = listOf(
+                    FilterMenu.UiState.Item(null, 0),
+                    FilterMenu.UiState.Item(Tool.CATEGORY_GOSPEL, 0),
+                    FilterMenu.UiState.Item(Tool.CATEGORY_ARTICLES, 0),
+                ),
+                query = mutableStateOf(Uuid.random().toString()),
+                selectedItem = null,
+                eventSink = {},
+            ),
+            languageFilter = FilterMenu.UiState(
+                menuExpanded = mutableStateOf(Random.nextBoolean()),
+                items = listOf(
+                    FilterMenu.UiState.Item(null, 0),
+                    FilterMenu.UiState.Item(Language(Locale.ENGLISH), 0),
+                    FilterMenu.UiState.Item(Language(Locale.FRENCH), 0),
+                ),
+                query = mutableStateOf(Uuid.random().toString()),
+                selectedItem = null,
+                eventSink = {},
+            )
         )
+        toolFiltersStateProducer.filters.value = filters
 
         presenter.test {
-            assertEquals(
-                listOf(
-                    FilterMenu.UiState.Item<String?>(Tool.CATEGORY_ARTICLES, 2),
-                    FilterMenu.UiState.Item<String?>(Tool.CATEGORY_GOSPEL, 1)
-                ),
-                expectMostRecentItem().filters.categoryFilter.items
-            )
+            assertEquals(filters, expectMostRecentItem().filters)
         }
     }
-
-    @Test
-    fun `State - filters - categoryFilter - items - uses selected language`() = testScope.runTest {
-        presenter.test {
-            awaitItem().filters.languageFilter.eventSink(FilterMenu.Event.SelectItem(Language(Locale.FRENCH)))
-            expectMostRecentItem()
-
-            verify { filteredToolsFlowProducer.getFlow(language = Locale.FRENCH) }
-        }
-    }
-    // endregion State.filters.categoryFilter.items
-
-    // region State.filters.languageFilter.items
-    @Test
-    fun `State - filters - languageFilter - items - no category`() = testScope.runTest {
-        val languages = listOf(Language(Locale.ENGLISH), Language(Locale.FRENCH))
-
-        presenter.test {
-            languagesFlow.value = languages
-            assertEquals(
-                listOf(
-                    FilterMenu.UiState.Item(null, 0),
-                    FilterMenu.UiState.Item(Language(Locale.ENGLISH), 0),
-                    FilterMenu.UiState.Item(Language(Locale.FRENCH), 0)
-                ),
-                expectMostRecentItem().filters.languageFilter.items
-            )
-        }
-
-        verifyAll {
-            languagesRepository.getLanguagesFlow()
-        }
-    }
-
-    @Test
-    fun `State - filters - languageFilter - items - for category`() = testScope.runTest {
-        val languages = listOf(Language(Locale.ENGLISH), Language(Locale.FRENCH))
-
-        presenter.test {
-            awaitItem().filters.categoryFilter.eventSink(FilterMenu.Event.SelectItem(Tool.CATEGORY_GOSPEL))
-
-            gospelLanguagesFlow.value = languages
-            assertEquals(
-                listOf(
-                    FilterMenu.UiState.Item(null, 0),
-                    FilterMenu.UiState.Item(Language(Locale.ENGLISH), 0),
-                    FilterMenu.UiState.Item(Language(Locale.FRENCH), 0)
-                ),
-                expectMostRecentItem().filters.languageFilter.items
-            )
-        }
-    }
-
-    @Test
-    fun `State - filters - languageFilter - items - include tool count`() = testScope.runTest {
-        val translationsFlow = MutableStateFlow(emptyList<Translation>())
-        every { translationsRepository.getTranslationsFlowForTools(setOf("tool1", "tool2")) } returns translationsFlow
-
-        presenter.test {
-            filteredToolsFlow.value = listOf(
-                randomTool("tool1", isHidden = false),
-                randomTool("tool2", isHidden = false),
-            )
-            translationsFlow.value = listOf(
-                randomTranslation("tool1", Locale.ENGLISH),
-                randomTranslation("tool1", Locale.FRENCH),
-                randomTranslation("tool2", Locale.ENGLISH, version = 1),
-                randomTranslation("tool2", Locale.ENGLISH, version = 2),
-            )
-            languagesFlow.value = listOf(Language(Locale.ENGLISH), Language(Locale.FRENCH))
-
-            assertEquals(
-                listOf(
-                    FilterMenu.UiState.Item(null, 0),
-                    FilterMenu.UiState.Item(Language(Locale.ENGLISH), 2),
-                    FilterMenu.UiState.Item(Language(Locale.FRENCH), 1)
-                ),
-                expectMostRecentItem().filters.languageFilter.items
-            )
-        }
-    }
-
-    @Test
-    fun `State - filters - languageFilter - items - filtered by query`() = testScope.runTest {
-        val languages = listOf(Language(Locale.ENGLISH), Language(Locale.FRENCH))
-        languagesFlow.value = languages
-
-        presenter.test {
-            expectMostRecentItem().filters.languageFilter.let {
-                it.menuExpanded.value = true
-                assertEquals(
-                    listOf(
-                        FilterMenu.UiState.Item(null, 0),
-                        FilterMenu.UiState.Item(Language(Locale.ENGLISH), 0),
-                        FilterMenu.UiState.Item(Language(Locale.FRENCH), 0),
-                    ),
-                    it.items
-                )
-                it.query.value = "french"
-            }
-
-            assertEquals(
-                listOf(
-                    FilterMenu.UiState.Item(null, 0),
-                    FilterMenu.UiState.Item(Language(Locale.FRENCH), 0)
-                ),
-                expectMostRecentItem().filters.languageFilter.items
-            )
-        }
-
-        verifyAll {
-            languagesRepository.getLanguagesFlow()
-        }
-    }
-    // endregion State.filters.languageFilter.items
-
-    // region State.filters.languageFilter.selectedItem
-    @Test
-    fun `State - filters - languageFilter - selectedItem - no language selected`() = testScope.runTest {
-        presenter.test {
-            assertNull(expectMostRecentItem().filters.languageFilter.selectedItem)
-        }
-
-        verify(inverse = true) { languagesRepository.findLanguageFlow(any()) }
-    }
-
-    @Test
-    fun `State - filters - languageFilter - selectedItem - language not found`() = testScope.runTest {
-        presenter.test {
-            awaitItem().filters.languageFilter.eventSink(FilterMenu.Event.SelectItem(Language(Locale.ENGLISH)))
-
-            assertNull(expectMostRecentItem().filters.languageFilter.selectedItem)
-        }
-
-        verify { languagesRepository.findLanguageFlow(Locale.ENGLISH) }
-    }
-
-    @Test
-    fun `State - filters - languageFilter - selectedItem - language selected`() = testScope.runTest {
-        val language = Language(Locale.ENGLISH)
-        every { languagesRepository.findLanguageFlow(Locale.ENGLISH) } returns flowOf(language)
-
-        presenter.test {
-            awaitItem().filters.languageFilter.eventSink(FilterMenu.Event.SelectItem(language))
-
-            assertEquals(language, expectMostRecentItem().filters.languageFilter.selectedItem)
-        }
-
-        verify { languagesRepository.findLanguageFlow(Locale.ENGLISH) }
-    }
-    // endregion State.filters.languageFilter.selectedItem
-
-    // region State.filters.languageFilter.menuExpanded
-    @Test
-    fun `State - filters - languageFilter - menuExpanded - resets query when set to false`() = testScope.runTest {
-        presenter.test {
-            val state = expectMostRecentItem().filters.languageFilter
-
-            state.menuExpanded.value = true
-            state.query.value = "test"
-            assertEquals("test", state.query.value)
-
-            state.menuExpanded.value = false
-            assertEquals("", state.query.value)
-
-            cancelAndIgnoreRemainingEvents()
-        }
-    }
-    // endregion FiltersEvent.ToggleLanguagesMenu
+    // endregion State.filters
 }
